#2800 fix IOT特性
